Adds stdlib.debug_mute
This commit adds a stdlib.debug_mute function that will print the command
being run, but not the output of the command, when Waffles is run in Debug
mode. The reason for this function is because there are times when it makes
sense to use stdlib.mute in non-debug mode.