Resource Refactor
This is a large commit that refactors the way resources work in Waffles.

Resources may now utilize the `stdlib.resource.process` function which reduces
a large amount of boilerplace code as well as enables pre- and post-steps to
be performed on each of the resources CRUD functions.

During the refactor, conditional quoting style was, once again, changed.

To ensure this refactor worked, new tests for almost all resource catagories
were introduced.

As well, documentation has been updated.