Template Support
This commit introduces basic templating support into Waffles.

The function `stdlib.render_template` takes a template file and
and an associative array / hash of variable as input. It then echoes
the rendered template which can be used as the contents of a file.

Only interpolation of `{{ varname }}` strings are supported at this time.