Large Refactor for 0.3
This commit is a major refactor of Waffles.

This repository can be considered Waffles "core". It contains the
core functionality of Waffles as a suite of Bash scripts. These
scripts can be sourced into any Bash script and then used to manage
resources. The init.sh script can also be source on the command-line
and the user can interactively execute the Waffles-based resources.

Functionality that has been removed from Waffles will appear in
separate repositories.